refactor(coins): extract pure data layer to feature_coins_core (#1111 slice 3)
Move the datasource interface (incl. its row-entity types), all 5
mappers, and all 5 repository implementations into feature_coins_core —
11 files, all pure Dart (dart:convert at most). The only genuinely
app-bound piece stays behind: coins_local_datasource_impl.dart (+stub),
the drift-backed binding to the monolithic AppDatabase. That impl is
now the single remaining seam between coins finance code and the app's
database — exactly the injection point the future CoinFinanceModule
will wire.

Old paths keep one-line re-export shims; 6 absolute-import consumers
switched to the package. docs-not-needed: shim-only edits under
app/lib/features/coins, no behavior change.
