+++ app/test/features/coins/coins_split_workflow_test.dart
@@ -2,8 +2,8 @@ import 'package:airo_app/core/database/app_database.dart';
 import 'package:airo_app/features/coins/application/use_cases/add_split_use_case.dart';
 import 'package:airo_app/features/coins/application/use_cases/create_group_use_case.dart';
 import 'package:airo_app/features/coins/data/datasources/coins_local_datasource_impl.dart';
-import 'package:airo_app/features/coins/data/mappers/group_mapper.dart';
-import 'package:airo_app/features/coins/data/repositories/group_repository_impl.dart';
+import 'package:feature_coins_core/src/data/mappers/group_mapper.dart';
+import 'package:feature_coins_core/src/data/repositories/group_repository_impl.dart';
 import 'package:feature_coins_core/src/entities/group.dart';
 import 'package:feature_coins_core/src/entities/group_member.dart';
 import 'package:feature_coins_core/src/entities/split_entry.dart';
