+++ packages/feature_coins_core/lib/feature_coins_core.dart
@@ -32,3 +32,14 @@ export 'src/services/finance_insight_service.dart';
 export 'src/services/finance_message_parser.dart';
 export 'src/services/quick_add_expense_parser.dart';
 export 'src/services/split_calculator.dart';
+export 'src/data/datasources/coins_local_datasource.dart';
+export 'src/data/mappers/account_mapper.dart';
+export 'src/data/mappers/budget_mapper.dart';
+export 'src/data/mappers/group_mapper.dart';
+export 'src/data/mappers/settlement_mapper.dart';
+export 'src/data/mappers/transaction_mapper.dart';
+export 'src/data/repositories/account_repository_impl.dart';
+export 'src/data/repositories/budget_repository_impl.dart';
+export 'src/data/repositories/group_repository_impl.dart';
+export 'src/data/repositories/settlement_repository_impl.dart';
+export 'src/data/repositories/transaction_repository_impl.dart';
